# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Felt API Doc",
        "description": "Felt API Documentation",
        "contact": {
            "email": "rejohnbd@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/web/admin/vehicle-types": {
            "get": {
                "tags": [
                    "getAllVehicleTypes"
                ],
                "summary": "Returns all vehicle type",
                "description": "",
                "operationId": "getAllVehicleTypes",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "addVehicleTypes"
                ],
                "operationId": "addVehicleTypes",
                "parameters": [
                    {
                        "name": "vehicle_type_name",
                        "in": "path",
                        "description": "Vehicle Type Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "vehicle_type_image",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}